+++ baker/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "extends": "../devTools/tsconfigs/tsconfig.base.json",
     "compilerOptions": {
+        "outDir": "./compiled/",
         "baseUrl": "../"
     },
     "references": [
